backend/pyproject.toml
[project]
name = "contract-intelligence-backend"
version = "0.1.0"
description = "Production-oriented contract intelligence backend"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.5.0",
  "python-multipart>=0.0.9",
  "langgraph>=0.2.40",
  "langchain-core>=0.3.0",
  "litellm>=1.52.0",
  "httpx>=0.27.0",
  "pypdf>=5.0.0",
  "python-docx>=1.1.2",
  "psycopg[binary]>=3.2.0",
  "langsmith>=0.1.0",
]

[project.optional-dependencies]
ocr = [
  "pytesseract>=0.3.10",
  "pdf2image>=1.17.0",
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
